+++ tools/python/boutiques/templates/template_docopt_descriptor.json
